+++ gen/functions.cpp
@@ -1139,11 +1139,7 @@ void DtoDefineFunction(FuncDeclaration *fd, bool linkageAvailableExternally) {
 
   // function attributes
   if (gABI->needsUnwindTables()) {
-#if LDC_LLVM_VER >= 1500
     func->setUWTableKind(llvm::UWTableKind::Default);
-#else
-    func->addFnAttr(LLAttribute::UWTable);
-#endif
   }
   if (opts::isAnySanitizerEnabled() &&
       !opts::functionIsInSanitizerBlacklist(fd)) {